feat: Implement Client and Supplier services with CRUD operations
- Added IClientService and ClientService for managing client data.
- Introduced ISupplierService and SupplierService for supplier management.
- Created ClientRequest and SupplierRequest DTOs for data transfer.
- Implemented API controllers for Client and Supplier with endpoints for adding, editing, retrieving, and listing.
- Updated IIncomeServices to include new invoice retrieval methods.
- Added GetInvoicesResponse and InvoiceItems classes for invoice data representation.
- Modified EF Core configurations and migrations to support new entities and relationships.

+++ Contract/Income/AddInvoicesRequest.cs
@@ -13,5 +13,5 @@ public class AddInvoiceRequest
 public class ClientInvoiceProductItem()
 {
       public Guid ProductId { get; set; }
-      public int Quantity { get; set; }
+      public Double Quantity { get; set; }
 }
